# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ClubHouse single-college club management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /colleges/{collegeId}/projects:
    get:
      operationId: getCollegeProjects
      tags: [colleges]
      summary: List projects belonging to a college
      parameters:
        - name: collegeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"

  # ── CLUBS ──────────────────────────────────────────────────────────────────────
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List public projects (discover)
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: open
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: Projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      security: [{ clerk: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project details
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project (owner/moderator)
      security: [{ clerk: [] }]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      security: [{ clerk: [] }]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/{projectId}/members:
    get:
      operationId: getProjectMembers
      tags: [projects]
      summary: List project members
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectMember"
  /projects/{projectId}/invites:
    post:
      operationId: inviteUserToProject
      tags: [projects]
      summary: Invite a user to a project
      security: [{ clerk: [] }]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInviteInput"
      responses:
        "201":
          description: User invited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectInvite"
  /projects/{projectId}/apply:
    post:
      operationId: applyToProject
      tags: [projects]
      summary: Apply to join a project
      security: [{ clerk: [] }]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectApplicationInput"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectApplication"
  /projects/{projectId}/applications:
    get:
      operationId: getProjectApplications
      tags: [projects]
      summary: List project applications (owner/moderator)
      security: [{ clerk: [] }]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectApplication"
  /projects/{projectId}/applications/{applicationId}:
    patch:
      operationId: updateProjectApplication
      tags: [projects]
      summary: Approve or reject project application
      security: [{ clerk: [] }]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: applicationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectApplication"
  /projects/{projectId}/messages:
    get:
      operationId: getProjectMessages
      tags: [projects]
      summary: Get project chat messages
      security: [{ clerk: [] }]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: before
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectMessage"
    post:
      operationId: sendProjectMessage
      tags: [projects]
      summary: Send a message in the project chat
      security: [{ clerk: [] }]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectMessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectMessage"
  /projects/{projectId}/events:
    get:
      operationId: getProjectEvents
      tags: [projects]
      summary: Get project calendar events
      security: [{ clerk: [] }]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectEvent"
    post:
      operationId: createProjectEvent
      tags: [projects]
      summary: Create a project calendar event
      security: [{ clerk: [] }]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectEventInput"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectEvent"
  /projects/{projectId}/events/{eventId}:
    delete:
      operationId: deleteProjectEvent
      tags: [projects]
      summary: Delete a project calendar event (owner/event creator/admin)
      security: [{ clerk: [] }]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── EVENTS (hackathons/workshops) ─────────────────────────────────────────────
...
components:
  securitySchemes:
    clerk:
      type: http
      scheme: bearer
  schemas:
...
    RequiredRole:
      type: object
      required: [id, role]
      properties:
        id:          { type: integer }
        role:        { type: string }
        description: { type: ["string", "null"] }
    Project:
      type: object
      required: [id, title, status, visibility, ownerId, createdAt]
      properties:
        id:            { type: integer }
        title:         { type: string }
        description:   { type: ["string", "null"] }
        techStack:     { type: ["string", "null"] }
        status:        { type: string, enum: [planning, active, completed, on_hold] }
        visibility:    { type: string, enum: [public, private, college_only] }
        ownerId:       { type: string }
        ownerName:     { type: string }
        collegeId:     { type: ["integer", "null"] }
        collegeName:   { type: ["string", "null"] }
        isJoint:       { type: boolean }
        partnerColleges:
          type: array
          items: { type: string }
        openForApplications: { type: boolean }
        requiredRoles:
          type: array
          items:
            $ref: "#/components/schemas/RequiredRole"
        memberCount:   { type: integer }
        createdAt:     { type: string, format: date-time }
    ProjectInput:
      type: object
      required: [title, visibility]
      properties:
        title:         { type: string }
        description:   { type: ["string", "null"] }
        techStack:     { type: ["string", "null"] }
        visibility:    { type: string, enum: [public, private, college_only] }
        collegeId:     { type: ["integer", "null"] }
        isJoint:       { type: boolean }
        openForApplications: { type: boolean }
        requiredRoles:
          type: array
          items:
            type: object
            required: [role]
            properties:
              role:        { type: string }
              description: { type: ["string", "null"] }
    ProjectUpdate:
      type: object
      properties:
        title:         { type: string }
        description:   { type: ["string", "null"] }
        techStack:     { type: ["string", "null"] }
        status:        { type: string, enum: [planning, active, completed, on_hold] }
        visibility:    { type: string, enum: [public, private, college_only] }
        openForApplications: { type: boolean }
        requiredRoles:
          type: array
          items:
            type: object
            required: [role]
            properties:
              role:        { type: string }
              description: { type: ["string", "null"] }
    ProjectMember:
      type: object
      required: [clerkId, name, role, joinedAt]
      properties:
        clerkId:   { type: string }
        name:      { type: string }
        avatarUrl: { type: ["string", "null"] }
        role:      { type: string }
        joinedAt:  { type: string, format: date-time }
    ProjectInviteInput:
      type: object
      required: [targetUserId]
      properties:
        targetUserId: { type: string }
    ProjectInvite:
      type: object
      required: [projectId, targetUserId, targetUserName, targetUserEmail, status, createdAt]
      properties:
        projectId:       { type: integer }
        targetUserId:    { type: string }
        targetUserName:  { type: string }
        targetUserEmail: { type: string }
        status:           { type: string, enum: [accepted] }
        createdAt:       { type: string, format: date-time }
    ProjectApplication:
      type: object
      required: [id, projectId, userId, status, createdAt]
      properties:
        id:          { type: integer }
        projectId:   { type: integer }
        userId:      { type: string }
        userName:    { type: string }
        userEmail:   { type: string }
        appliedRole: { type: ["string", "null"] }
        message:     { type: ["string", "null"] }
        status:      { type: string, enum: [pending, approved, rejected] }
        createdAt:   { type: string, format: date-time }
    ProjectApplicationInput:
      type: object
      properties:
        appliedRole: { type: ["string", "null"] }
        message:     { type: ["string", "null"] }
    ProjectMessage:
      type: object
      required: [id, projectId, userId, content, createdAt]
      properties:
        id:        { type: integer }
        projectId: { type: integer }
        userId:    { type: string }
        userName:  { type: string }
        avatarUrl: { type: ["string", "null"] }
        content:   { type: string }
        createdAt: { type: string, format: date-time }
    ProjectMessageInput:
      type: object
      required: [content]
      properties:
        content: { type: string }
    ProjectEvent:
      type: object
      required: [id, projectId, title, scheduledAt, createdAt]
      properties:
        id:          { type: integer }
        projectId:   { type: integer }
        title:       { type: string }
        description: { type: ["string", "null"] }
        meetLink:    { type: ["string", "null"] }
        scheduledAt: { type: string, format: date-time }
        createdByName: { type: string }
        createdBy:    { type: string }
        createdAt:   { type: string, format: date-time }
    ProjectEventInput:
      type: object
      required: [title, scheduledAt]
      properties:
        title:       { type: string }
        description: { type: ["string", "null"] }
        meetLink:    { type: ["string", "null"] }
        scheduledAt: { type: string, format: date-time }

    # ── EVENTS ───────────────────────────────────────────────────────────────────
...
    StatusUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [approved, rejected] }
...